Move utility fn into vertex file
