+++ sdks/python/apache_beam/ml/inference/vertex_ai_inference_test.py
@@ -1,0 +1,41 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# pytype: skip-file
+
+import unittest
+
+from apache_beam.ml.inference.vertex_ai_inference import retry_on_gcp_client_error
+
+try:
+  from google.api_core.exceptions import TooManyRequests
+except ImportError:
+  TooManyRequests = None  # type: ignore
+
+
+@unittest.skipIf(TooManyRequests is None, 'GCP dependencies are not installed')
+class RetryOnClientErrorTest(unittest.TestCase):
+  def test_retry_on_client_error_positive(self):
+    e = TooManyRequests(message="fake service rate limiting")
+    self.assertTrue(retry_on_gcp_client_error(e))
+
+  def test_retry_on_client_error_negative(self):
+    e = ValueError()
+    self.assertFalse(retry_on_gcp_client_error(e))
+
+
+if __name__ == '__main__':
+  unittest.main()
